[Added]: A general applicable aspect that also handles the pipeline-side. [Missing]: Aspect that handles component-side
